requirements.txt
python-escpos==3.0a4
paho-mqtt
aiosmtpd
bs4
w3lib
emoji
regex
lxml
tinycss
cssselect